@@ -1,29 +1,32 @@
 module HalogenGhostCore
 
-go 1.19
+go 1.22.0
+
+toolchain go1.23.1
 
 require (
-	github.com/aws/aws-sdk-go v1.50.10
-	github.com/getsentry/sentry-go v0.26.0
+	github.com/aws/aws-sdk-go v1.55.5
+	github.com/getsentry/sentry-go v0.28.1
 	github.com/go-co-op/gocron v1.37.0
 	github.com/go-redis/redis/v8 v8.11.5
-	github.com/go-sql-driver/mysql v1.7.1
+	github.com/go-sql-driver/mysql v1.8.1
 	github.com/gorilla/mux v1.8.1
-	github.com/hashicorp/consul/api v1.27.0
-	github.com/jmoiron/sqlx v1.3.5
-	github.com/rabbitmq/amqp091-go v1.9.0
-	golang.org/x/exp v0.0.0-20240119083558-1b970713d09a
+	github.com/hashicorp/consul/api v1.29.4
+	github.com/jmoiron/sqlx v1.4.0
+	github.com/lesismal/arpc v1.2.16
 )
 
 require (
+	filippo.io/edwards25519 v1.1.0 // indirect
 	github.com/armon/go-metrics v0.4.1 // indirect
-	github.com/cespare/xxhash/v2 v2.2.0 // indirect
+	github.com/benbjohnson/clock v1.3.0 // indirect
+	github.com/cespare/xxhash/v2 v2.3.0 // indirect
 	github.com/dgryski/go-rendezvous v0.0.0-20200823014737-9f7001d12a5f // indirect
-	github.com/fatih/color v1.16.0 // indirect
+	github.com/fatih/color v1.17.0 // indirect
 	github.com/google/uuid v1.6.0 // indirect
 	github.com/hashicorp/errwrap v1.1.0 // indirect
 	github.com/hashicorp/go-cleanhttp v0.5.2 // indirect
-	github.com/hashicorp/go-hclog v1.6.2 // indirect
+	github.com/hashicorp/go-hclog v1.6.3 // indirect
 	github.com/hashicorp/go-immutable-radix v1.3.1 // indirect
 	github.com/hashicorp/go-multierror v1.1.1 // indirect
 	github.com/hashicorp/go-rootcerts v1.0.2 // indirect
@@ -36,6 +39,8 @@ require (
 	github.com/mitchellh/mapstructure v1.5.0 // indirect
 	github.com/robfig/cron/v3 v3.0.1 // indirect
 	go.uber.org/atomic v1.11.0 // indirect
-	golang.org/x/sys v0.16.0 // indirect
-	golang.org/x/text v0.14.0 // indirect
+	go.uber.org/ratelimit v0.3.1 // indirect
+	golang.org/x/exp v0.0.0-20240909161429-701f63a606c0 // indirect
+	golang.org/x/sys v0.25.0 // indirect
+	golang.org/x/text v0.18.0 // indirect
 )
