@@ -9,38 +9,33 @@ import (
 	"io"
 	"net/http"
 	"os"
-	"strconv"
-	"strings"
 )
 
 func AccountCommentDelete(resp http.ResponseWriter, req *http.Request, conf *core.GlobalConfig) {
-	IPAddr := req.Header.Get("CF-Connecting-IP")
-	if IPAddr == "" {
-		IPAddr = req.Header.Get("X-Real-IP")
-	}
-	if IPAddr == "" {
-		IPAddr = strings.Split(req.RemoteAddr, ":")[0]
-	}
+	IPAddr := ipOf(req)
 	vars := gorilla.Vars(req)
 	logger := core.Logger{Output: os.Stderr}
+	connector := connectors.NewConnector(req.URL.Query().Has("json"))
+	defer func() { _, _ = io.WriteString(resp, connector.Output()) }()
 	config, err := conf.LoadById(vars["gdps"])
 	if logger.Should(err) != nil {
+		connector.Error("-1", "Not Found")
 		return
 	}
 	if core.CheckIPBan(IPAddr, config) {
+		connector.Error("-1", "Banned")
 		return
 	}
-	//Get:=req.URL.Query()
 	Post := ReadPost(req)
 	if core.CheckGDAuth(Post) && Post.Get("commentID") != "" {
 		db := &core.MySQLConn{}
-		defer db.CloseDB()
 		if logger.Should(db.ConnectBlob(config)) != nil {
+			serverError(connector)
 			return
 		}
 		xacc := core.CAccount{DB: db}
 		if !xacc.PerformGJPAuth(Post, IPAddr) {
-			io.WriteString(resp, "-1")
+			connector.Error("-1", "Invalid Credentials")
 			return
 		}
 		cc := core.CComment{DB: db}
@@ -48,37 +43,36 @@ func AccountCommentDelete(resp http.ResponseWriter, req *http.Request, conf *cor
 		core.TryInt(&id, Post.Get("commentID"))
 		cc.DeleteAccComment(id, xacc.Uid)
 		core.OnPost(db, conf, config)
-		io.WriteString(resp, "1")
+		connector.Success("Comment deleted")
 	} else {
-		io.WriteString(resp, "-1")
+		connector.Error("-1", "Bad Request")
 	}
 }
 
 func AccountCommentGet(resp http.ResponseWriter, req *http.Request, conf *core.GlobalConfig) {
-	IPAddr := req.Header.Get("CF-Connecting-IP")
-	if IPAddr == "" {
-		IPAddr = req.Header.Get("X-Real-IP")
-	}
-	if IPAddr == "" {
-		IPAddr = strings.Split(req.RemoteAddr, ":")[0]
-	}
+	IPAddr := ipOf(req)
 	vars := gorilla.Vars(req)
 	logger := core.Logger{Output: os.Stderr}
+	connector := connectors.NewConnector(req.URL.Query().Has("json"))
+	defer func() { _, _ = io.WriteString(resp, connector.Output()) }()
 	config, err := conf.LoadById(vars["gdps"])
 	if logger.Should(err) != nil {
+		connector.Error("-1", "Not Found")
 		return
 	}
 	if core.CheckIPBan(IPAddr, config) {
+		connector.Error("-1", "Banned")
 		return
 	}
-	//Get:=req.URL.Query()
+
 	Post := ReadPost(req)
 	if Post.Get("accountID") != "" {
 		page := 0
 		core.TryInt(&page, Post.Get("page"))
 		db := &core.MySQLConn{}
-		defer db.CloseDB()
+
 		if logger.Should(db.ConnectBlob(config)) != nil {
+			serverError(connector)
 			return
 		}
 		var uid int
@@ -90,33 +84,21 @@ func AccountCommentGet(resp http.ResponseWriter, req *http.Request, conf *core.G
 		core.TryInt(&uid, xuid)
 		cc := core.CComment{DB: db}
 		comments := cc.GetAllAccComments(uid, page)
-		if len(comments) == 0 {
-			io.WriteString(resp, "#0:0:0")
-		} else {
-			output := ""
-			for _, comm := range comments {
-				output += connectors.GetAccountComment(comm)
-			}
-			io.WriteString(resp, output[:len(output)-1]+"#"+strconv.Itoa(cc.CountAccComments(uid))+":"+strconv.Itoa(page*10)+":10")
-		}
-
+		connector.Comment_AccountGet(comments, cc.CountAccComments(uid), page)
 	} else {
-		io.WriteString(resp, "-1")
+		connector.Error("-1", "Bad Request")
 	}
 }
 
 func AccountCommentUpload(resp http.ResponseWriter, req *http.Request, conf *core.GlobalConfig) {
-	IPAddr := req.Header.Get("CF-Connecting-IP")
-	if IPAddr == "" {
-		IPAddr = req.Header.Get("X-Real-IP")
-	}
-	if IPAddr == "" {
-		IPAddr = strings.Split(req.RemoteAddr, ":")[0]
-	}
+	IPAddr := ipOf(req)
 	vars := gorilla.Vars(req)
 	logger := core.Logger{Output: os.Stderr}
+	connector := connectors.NewConnector(req.URL.Query().Has("json"))
+	defer func() { _, _ = io.WriteString(resp, connector.Output()) }()
 	config, err := conf.LoadById(vars["gdps"])
 	if logger.Should(err) != nil {
+		connector.Error("-1", "Not Found")
 		return
 	}
 
@@ -125,66 +107,67 @@ func AccountCommentUpload(resp http.ResponseWriter, req *http.Request, conf *cor
 	}
 
 	if core.CheckIPBan(IPAddr, config) {
+		connector.Error("-1", "Banned")
 		return
 	}
-	//Get:=req.URL.Query()
+
 	Post := ReadPost(req)
 	if core.CheckGDAuth(Post) && Post.Get("comment") != "" {
 		db := &core.MySQLConn{}
-		defer db.CloseDB()
+
 		if logger.Should(db.ConnectBlob(config)) != nil {
+			serverError(connector)
 			return
 		}
 		xacc := core.CAccount{DB: db}
 		if !xacc.PerformGJPAuth(Post, IPAddr) {
-			io.WriteString(resp, "-1")
+			connector.Error("-1", "Invalid Credentials")
 			return
 		}
 		comment := core.ClearGDRequest(Post.Get("comment"))
 		cc := core.CComment{DB: db, Uid: xacc.Uid, Comment: comment}
 		protect := core.CProtect{DB: db, DisableProtection: config.SecurityConfig.DisableProtection}
-		c := "-1"
 		if !core.OnPost(db, conf, config) {
-			io.WriteString(resp, "-1")
+			connector.Error("-1", "Post limits exceeded")
 			return
 		}
 		if protect.DetectPosts(xacc.Uid) && cc.PostAccComment() {
-			c = "1"
+			connector.Success("Comment posted")
+			return
 		}
-		io.WriteString(resp, c)
+		serverError(connector)
 	} else {
-		io.WriteString(resp, "-1")
+		connector.Error("-1", "Bad Request")
 	}
 }
 
 func CommentDelete(resp http.ResponseWriter, req *http.Request, conf *core.GlobalConfig) {
-	IPAddr := req.Header.Get("CF-Connecting-IP")
-	if IPAddr == "" {
-		IPAddr = req.Header.Get("X-Real-IP")
-	}
-	if IPAddr == "" {
-		IPAddr = strings.Split(req.RemoteAddr, ":")[0]
-	}
+	IPAddr := ipOf(req)
 	vars := gorilla.Vars(req)
 	logger := core.Logger{Output: os.Stderr}
+	connector := connectors.NewConnector(req.URL.Query().Has("json"))
+	defer func() { _, _ = io.WriteString(resp, connector.Output()) }()
 	config, err := conf.LoadById(vars["gdps"])
 	if logger.Should(err) != nil {
+		connector.Error("-1", "Not Found")
 		return
 	}
 	if core.CheckIPBan(IPAddr, config) {
+		connector.Error("-1", "Banned")
 		return
 	}
-	//Get:=req.URL.Query()
+
 	Post := ReadPost(req)
 	if core.CheckGDAuth(Post) && Post.Get("commentID") != "" && Post.Get("levelID") != "" {
 		db := &core.MySQLConn{}
-		defer db.CloseDB()
+
 		if logger.Should(db.ConnectBlob(config)) != nil {
+			serverError(connector)
 			return
 		}
 		xacc := core.CAccount{DB: db}
 		if !xacc.PerformGJPAuth(Post, IPAddr) {
-			io.WriteString(resp, "-1")
+			connector.Error("-1", "Invalid Credentials")
 			return
 		}
 		cc := core.CComment{DB: db}
@@ -209,30 +192,28 @@ func CommentDelete(resp http.ResponseWriter, req *http.Request, conf *core.Globa
 			}
 		}
 		core.OnComment(db, conf, config)
-		io.WriteString(resp, "1")
+		connector.Success("Comment deleted")
 	} else {
-		io.WriteString(resp, "-1")
+		connector.Error("-1", "Bad Request")
 	}
 }
 
 func CommentGet(resp http.ResponseWriter, req *http.Request, conf *core.GlobalConfig) {
-	IPAddr := req.Header.Get("CF-Connecting-IP")
-	if IPAddr == "" {
-		IPAddr = req.Header.Get("X-Real-IP")
-	}
-	if IPAddr == "" {
-		IPAddr = strings.Split(req.RemoteAddr, ":")[0]
-	}
+	IPAddr := ipOf(req)
 	vars := gorilla.Vars(req)
 	logger := core.Logger{Output: os.Stderr}
+	connector := connectors.NewConnector(req.URL.Query().Has("json"))
+	defer func() { _, _ = io.WriteString(resp, connector.Output()) }()
 	config, err := conf.LoadById(vars["gdps"])
 	if logger.Should(err) != nil {
+		connector.Error("-1", "Not Found")
 		return
 	}
 	if core.CheckIPBan(IPAddr, config) {
+		connector.Error("-1", "Banned")
 		return
 	}
-	//Get:=req.URL.Query()
+
 	Post := ReadPost(req)
 	if Post.Get("levelID") != "" {
 		page := 0
@@ -242,47 +223,37 @@ func CommentGet(resp http.ResponseWriter, req *http.Request, conf *core.GlobalCo
 			mode = true
 		}
 		db := &core.MySQLConn{}
-		defer db.CloseDB()
+
 		if logger.Should(db.ConnectBlob(config)) != nil {
+			serverError(connector)
 			return
 		}
 		var lvlId int
 		core.TryInt(&lvlId, Post.Get("levelID"))
 		cc := core.CComment{DB: db}
 		comments := cc.GetAllLevelComments(lvlId, page, mode)
-		if len(comments) == 0 {
-			io.WriteString(resp, "#0:0:0")
-		} else {
-			output := ""
-			for _, comm := range comments {
-				output += connectors.GetLevelComment(comm)
-			}
-			io.WriteString(resp, output[:core.MaxInt(len(output)-1, 0)]+"#"+strconv.Itoa(cc.CountLevelComments(lvlId))+":"+strconv.Itoa(page*10)+":10")
-		}
-
+		connector.Comment_LevelGet(comments, cc.CountLevelComments(lvlId), page)
 	} else {
-		io.WriteString(resp, "-1")
+		connector.Error("-1", "Bad Request")
 	}
 }
 
 func CommentGetHistory(resp http.ResponseWriter, req *http.Request, conf *core.GlobalConfig) {
-	IPAddr := req.Header.Get("CF-Connecting-IP")
-	if IPAddr == "" {
-		IPAddr = req.Header.Get("X-Real-IP")
-	}
-	if IPAddr == "" {
-		IPAddr = strings.Split(req.RemoteAddr, ":")[0]
-	}
+	IPAddr := ipOf(req)
 	vars := gorilla.Vars(req)
 	logger := core.Logger{Output: os.Stderr}
+	connector := connectors.NewConnector(req.URL.Query().Has("json"))
+	defer func() { _, _ = io.WriteString(resp, connector.Output()) }()
 	config, err := conf.LoadById(vars["gdps"])
 	if logger.Should(err) != nil {
+		connector.Error("-1", "Not Found")
 		return
 	}
 	if core.CheckIPBan(IPAddr, config) {
+		connector.Error("-1", "Banned")
 		return
 	}
-	//Get:=req.URL.Query()
+
 	Post := ReadPost(req)
 	if Post.Get("userID") != "" {
 		page := 0
@@ -292,14 +263,15 @@ func CommentGetHistory(resp http.ResponseWriter, req *http.Request, conf *core.G
 			mode = true
 		}
 		db := &core.MySQLConn{}
-		defer db.CloseDB()
+
 		if logger.Should(db.ConnectBlob(config)) != nil {
+			serverError(connector)
 			return
 		}
 		acc := core.CAccount{DB: db}
 		core.TryInt(&acc.Uid, Post.Get("userID"))
 		if !acc.Exists(acc.Uid) {
-			io.WriteString(resp, "-1")
+			connector.Error("-1", "Invalid Credentials")
 			return
 		}
 		acc.LoadAuth(core.CAUTH_UID)
@@ -308,49 +280,39 @@ func CommentGetHistory(resp http.ResponseWriter, req *http.Request, conf *core.G
 		role := acc.GetRoleObj(false)
 		cc := core.CComment{DB: db}
 		comments := cc.GetAllCommentsHistory(acc.Uid, page, mode)
-		if len(comments) == 0 {
-			io.WriteString(resp, "#0:0:0")
-		} else {
-			output := ""
-			for _, comm := range comments {
-				output += connectors.GetCommentHistory(comm, acc, role)
-			}
-			io.WriteString(resp, output[:len(output)-1]+"#"+strconv.Itoa(cc.CountCommentHistory(acc.Uid))+":"+strconv.Itoa(page*10)+":10")
-		}
-
+		connector.Comment_HistoryGet(comments, acc, role, cc.CountCommentHistory(acc.Uid), page)
 	} else {
-		io.WriteString(resp, "-1")
+		connector.Error("-1", "Bad Request")
 	}
 }
 
 func CommentUpload(resp http.ResponseWriter, req *http.Request, conf *core.GlobalConfig) {
-	IPAddr := req.Header.Get("CF-Connecting-IP")
-	if IPAddr == "" {
-		IPAddr = req.Header.Get("X-Real-IP")
-	}
-	if IPAddr == "" {
-		IPAddr = strings.Split(req.RemoteAddr, ":")[0]
-	}
+	IPAddr := ipOf(req)
 	vars := gorilla.Vars(req)
 	logger := core.Logger{Output: os.Stderr}
+	connector := connectors.NewConnector(req.URL.Query().Has("json"))
+	defer func() { _, _ = io.WriteString(resp, connector.Output()) }()
 	config, err := conf.LoadById(vars["gdps"])
 	if logger.Should(err) != nil {
+		connector.Error("-1", "Not Found")
 		return
 	}
 	if core.CheckIPBan(IPAddr, config) {
+		connector.Error("-1", "Banned")
 		return
 	}
-	//Get:=req.URL.Query()
+
 	Post := ReadPost(req)
 	if core.CheckGDAuth(Post) && Post.Get("comment") != "" && Post.Get("levelID") != "" {
 		db := &core.MySQLConn{}
-		defer db.CloseDB()
+
 		if logger.Should(db.ConnectBlob(config)) != nil {
+			serverError(connector)
 			return
 		}
 		xacc := core.CAccount{DB: db}
 		if !xacc.PerformGJPAuth(Post, IPAddr) {
-			io.WriteString(resp, "-1")
+			connector.Error("-1", "Invalid Credentials")
 			return
 		}
 		comment := core.ClearGDRequest(Post.Get("comment"))
@@ -363,26 +325,26 @@ func CommentUpload(resp http.ResponseWriter, req *http.Request, conf *core.Globa
 			// CLevelList
 			list := core.CLevelList{DB: db, ID: cl.Id * -1}
 			if !list.Exists(list.ID) {
-				io.WriteString(resp, "-1")
+				connector.Error("-1", "Invalid LevelList ID")
 				return
 			}
 			cc := core.CComment{DB: db, Uid: xacc.Uid, LvlId: cl.Id, Comment: comment, Percent: percent}
 			protect := core.CProtect{DB: db, DisableProtection: config.SecurityConfig.DisableProtection}
 			if !core.OnComment(db, conf, config) {
-				io.WriteString(resp, "-1")
+				connector.Error("-1", "Comment limit exceeded")
 				return
 			}
 			if protect.DetectComments(xacc.Uid) && cc.PostLevelComment() {
-				io.WriteString(resp, "1")
+				connector.Success("Comment posted")
 			} else {
-				io.WriteString(resp, "-1")
+				serverError(connector)
 			}
 			return
 		}
 
 		// Next
 		if !cl.Exists(cl.Id) {
-			io.WriteString(resp, "-1")
+			connector.Error("-1", "Invalid Level ID")
 			return
 		}
 		acc := core.CAccount{DB: db, Uid: xacc.Uid}
@@ -398,40 +360,39 @@ func CommentUpload(resp http.ResponseWriter, req *http.Request, conf *core.Globa
 				cmdRes := core.InvokeCommands(db, cl, acc, modComment, isOwned, role)
 				switch cmdRes {
 				case "err":
-					io.WriteString(resp, "-1")
+					connector.Error("-1", cmdRes)
 				case "ok":
-					io.WriteString(resp, "1")
+					connector.Success("Command executed")
 				default:
-					io.WriteString(resp, "temp_1_"+cmdRes)
+					connector.Error("-1", cmdRes)
 				}
 			} else {
 				cc := core.CComment{DB: db, Uid: xacc.Uid, LvlId: cl.Id, Comment: comment, Percent: percent}
 				protect := core.CProtect{DB: db, DisableProtection: config.SecurityConfig.DisableProtection}
 				if !core.OnComment(db, conf, config) {
-					io.WriteString(resp, "-1")
+					connector.Error("-1", "Comment limit exceeded")
 					return
 				}
 				if protect.DetectComments(xacc.Uid) && cc.PostLevelComment() {
-					io.WriteString(resp, "1")
+					connector.Success("Comment posted")
 				} else {
-					io.WriteString(resp, "-1")
+					serverError(connector)
 				}
 			}
-
 		} else {
 			cc := core.CComment{DB: db, Uid: xacc.Uid, LvlId: cl.Id, Comment: comment, Percent: percent}
 			protect := core.CProtect{DB: db, DisableProtection: config.SecurityConfig.DisableProtection}
 			if !core.OnComment(db, conf, config) {
-				io.WriteString(resp, "-1")
+				connector.Error("-1", "Comment limit exceeded")
 				return
 			}
 			if protect.DetectComments(xacc.Uid) && cc.PostLevelComment() {
-				io.WriteString(resp, "1")
+				connector.Success("Comment posted")
 			} else {
-				io.WriteString(resp, "-1")
+				serverError(connector)
 			}
 		}
 	} else {
-		io.WriteString(resp, "-1")
+		connector.Error("-1", "Bad Request")
 	}
 }
