@@ -8,33 +8,31 @@ import (
 	"io"
 	"net/http"
 	"os"
-	"strings"
 	"time"
 )
 
 func GetChallenges(resp http.ResponseWriter, req *http.Request, conf *core.GlobalConfig) {
-	IPAddr := req.Header.Get("CF-Connecting-IP")
-	if IPAddr == "" {
-		IPAddr = req.Header.Get("X-Real-IP")
-	}
-	if IPAddr == "" {
-		IPAddr = strings.Split(req.RemoteAddr, ":")[0]
-	}
+	IPAddr := ipOf(req)
 	vars := gorilla.Vars(req)
 	logger := core.Logger{Output: os.Stderr}
+	connector := connectors.NewConnector(req.URL.Query().Has("json"))
+	defer func() { _, _ = io.WriteString(resp, connector.Output()) }()
 	config, err := conf.LoadById(vars["gdps"])
 	if logger.Should(err) != nil {
+		connector.Error("-1", "Not Found")
 		return
 	}
 	if core.CheckIPBan(IPAddr, config) {
+		connector.Error("-1", "Banned")
 		return
 	}
-	//Get:=req.URL.Query()
+
 	Post := ReadPost(req)
 	if Post.Get("chk") != "" && Post.Get("udid") != "" {
 		db := &core.MySQLConn{}
-		defer db.CloseDB()
+
 		if logger.Should(db.ConnectBlob(config)) != nil {
+			serverError(connector)
 			return
 		}
 		cq := core.CQuests{DB: db}
@@ -43,43 +41,42 @@ func GetChallenges(resp http.ResponseWriter, req *http.Request, conf *core.Globa
 			chk := core.DoXOR(string(chalk), "19847")
 			var uid int
 			core.TryInt(&uid, Post.Get("accountID"))
-			io.WriteString(resp, connectors.ChallengesOutput(cq, uid, chk, Post.Get("udid")))
+			connector.Rewards_ChallengesOutput(cq, uid, chk, Post.Get("udid"))
 		} else {
-			io.WriteString(resp, "-2")
+			connector.Error("-2", "Challenge not found")
 		}
 	} else {
-		io.WriteString(resp, "-1")
+		connector.Error("-1", "Bad Request")
 	}
 }
 
 func GetRewards(resp http.ResponseWriter, req *http.Request, conf *core.GlobalConfig) {
-	IPAddr := req.Header.Get("CF-Connecting-IP")
-	if IPAddr == "" {
-		IPAddr = req.Header.Get("X-Real-IP")
-	}
-	if IPAddr == "" {
-		IPAddr = strings.Split(req.RemoteAddr, ":")[0]
-	}
+	IPAddr := ipOf(req)
 	vars := gorilla.Vars(req)
 	logger := core.Logger{Output: os.Stderr}
+	connector := connectors.NewConnector(req.URL.Query().Has("json"))
+	defer func() { _, _ = io.WriteString(resp, connector.Output()) }()
 	config, err := conf.LoadById(vars["gdps"])
 	if logger.Should(err) != nil {
+		connector.Error("-1", "Not Found")
 		return
 	}
 	if core.CheckIPBan(IPAddr, config) {
+		connector.Error("-1", "Banned")
 		return
 	}
-	//Get:=req.URL.Query()
+
 	Post := ReadPost(req)
 	if core.CheckGDAuth(Post) && len(Post.Get("chk")) > 5 && Post.Get("udid") != "" {
 		db := &core.MySQLConn{}
-		defer db.CloseDB()
+
 		if logger.Should(db.ConnectBlob(config)) != nil {
+			serverError(connector)
 			return
 		}
 		xacc := core.CAccount{DB: db}
 		if !xacc.PerformGJPAuth(Post, IPAddr) {
-			io.WriteString(resp, "-1")
+			connector.Error("-1", "Invalid Credentials")
 			return
 		}
 		var chestType int
@@ -99,7 +96,7 @@ func GetRewards(resp http.ResponseWriter, req *http.Request, conf *core.GlobalCo
 				xacc.PushChests()
 				chestSmallLeft = config.ChestConfig.ChestSmallWait
 			} else {
-				io.WriteString(resp, "-1")
+				connector.Error("-1", "Small chest is not ready yet")
 				return
 			}
 		case 2:
@@ -109,12 +106,12 @@ func GetRewards(resp http.ResponseWriter, req *http.Request, conf *core.GlobalCo
 				xacc.PushChests()
 				chestBigLeft = config.ChestConfig.ChestBigWait
 			} else {
-				io.WriteString(resp, "-1")
+				connector.Error("-1", "Big chest is not ready yet")
 				return
 			}
 		}
-		io.WriteString(resp, connectors.ChestOutput(xacc, config, Post.Get("udid"), chk, chestSmallLeft, chestBigLeft, chestType))
+		connector.Rewards_ChestOutput(xacc, config, Post.Get("udid"), chk, chestSmallLeft, chestBigLeft, chestType)
 	} else {
-		io.WriteString(resp, "-1")
+		connector.Error("-1", "Bad Request")
 	}
 }
