@@ -16,13 +16,13 @@ import (
 	"github.com/aws/aws-sdk-go/service/s3/s3manager"
 	"github.com/getsentry/sentry-go"
 	"github.com/go-redis/redis/v8"
-	"golang.org/x/exp/slices"
 	"html"
 	"io"
 	"math/rand"
 	"net/http"
 	"net/url"
 	"os"
+	"slices"
 	"strconv"
 	"strings"
 	"time"
@@ -292,6 +292,16 @@ func ArrTranslate(arr []int) []string {
 	return vs
 }
 
+func ArrTranslateToInt(arr []string) []int {
+	var vs []int
+	for _, l := range arr {
+		if v, err := strconv.Atoi(l); err == nil {
+			vs = append(vs, v)
+		}
+	}
+	return vs
+}
+
 func QuickComma(str string) string {
 	return strings.Join(ArrTranslate(Decompose(CleanDoubles(str, ","), ",")), ",")
 }
@@ -547,4 +557,14 @@ func (gm *GoMetrics) DumpJSON() string {
 		Total: gm.lastTime.Sub(gm.startTime).Milliseconds(),
 	})
 	return string(data)
+}
+
+func Clamp(val, min, max int) int {
+	if val < min {
+		return min
+	}
+	if val > max {
+		return max
+	}
+	return val
 }
