@@ -1,6 +1,3 @@
-//go:build (linux && amd64) || (windows && amd64)
-// +build linux,amd64 windows,amd64
-
 package main
 
 import (
@@ -13,6 +10,8 @@ import (
 	"os"
 	"time"
 )
+
+var restartAttempts = 3
 
 func main() {
 	// Start Sentry so I can sleep well
@@ -38,6 +37,10 @@ func main() {
 
 	core.DBTunnel, err = sqlx.Connect("mysql", DB_USER+":"+DB_PASS+"@tcp("+DB_HOST+":3306)/")
 	if err != nil {
+		if restartAttempts == 0 {
+			os.Exit(1)
+		}
+		restartAttempts--
 		log.Println("Error while connecting to " + DB_USER + "@localhost: " + err.Error())
 		time.Sleep(10 * time.Second)
 		main()
@@ -49,7 +52,7 @@ func main() {
 		},
 		Config: config,
 	}
-
+	api.InitCache(config, 16)
 	core.PrepareElection(config)
 	defer core.StepDown()
 
@@ -58,22 +61,16 @@ func main() {
 
 func GenGConfig() core.GlobalConfig {
 	return core.GlobalConfig{
-		EnvOrDefault("MASTER_KEY", "3XTR4OrD1nArY_K3Y_1907"),
-		EnvOrDefault("API_ENDPOINT", "https://api.fruitspace.one/sched/gd/api"),
-		"stdout",
-		"null",
-		false,
-		EnvOrDefault("REDIS_HOST", "localhost"),
-		EnvOrDefault("REDIS_PORT", "6379"),
-		EnvOrDefault("REDIS_PASSWORD", ""),
-		7,
-		EnvOrDefault("SAVE_PATH", "./"),
-
-		map[string]string{
-			"rabbitmq_host":     "auto",
-			"rabbitmq_user":     "m41dss",
-			"rabbitmq_password": EnvOrDefault("RABBITMQ_PASSWORD", "None"),
-		},
+		MasterKey:      EnvOrDefault("MASTER_KEY", "3XTR4OrD1nArY_K3Y_1907"),
+		ApiEndpoint:    EnvOrDefault("API_ENDPOINT", "http://127.0.0.1:6000/sched/gd/api"),
+		LogConnector:   "stdout",
+		LogEndpoint:    "null",
+		RedisHost:      EnvOrDefault("REDIS_HOST", "localhost"),
+		RedisPort:      EnvOrDefault("REDIS_PORT", "6379"),
+		RedisPassword:  EnvOrDefault("REDIS_PASSWORD", ""),
+		RedisDB:        7,
+		SavePath:       EnvOrDefault("SAVE_PATH", "./"),
+		ModuleSettings: map[string]string{},
 	}
 }
 
